lib: cgroup_bw: Implement replenish timer.
We replenish the cgroup's budget every 100ms (nperiod) interval. Only
subroot cgroups (their levels == 1) distribute the budget to their
descendants; we only replenish the budget for the subroot cgroup with a
limited quota. The underused budget can be accumulated by the burst
specified. On the other hand, the overused budget will be charged over the
intervals.

Signed-off-by: Changwoo Min <changwoo@igalia.com>